// src/app/models/map.model.ts
import { ICharacter } from "./character.model";

export interface IScene {
    description: string;
    backgroundColor: string;
    traversable: boolean;
}

export interface ILocation {
    row: number;
    col: number;
    character: ICharacter | null;
    scene: IScene;
}

export interface IMap {
    // rows: number;
    // cols: number;
    locations: ILocation[][]
}